Add support for Python 3.6
Switch to using zope-testrunner instead of setup.py test because of
the namespace package issues.

Enable pip caching on travis.

Stop testing pypy3 on travis since it implements Python 3.2 and is not
supported by pip anymore. There is a 3.3-alpha available on travis,
but that's a dead end and not developed anymore since 3.5-beta is
out (but not available on Travis AFAIK).

Fixes #6.