Configure with future Python 3.13.
This is a testing PR for https://github.com/zopefoundation/meta/pull/218
It introduces Python hyphen ranges in `tests.yml` for the future Python 3.13.
Manually added `3.11.0 - 3.11.2` to see the effect for a current Python.

To make the code differences smaller, this branch is based on https://github.com/zopefoundation/zope.exceptions/pull/30 which updated to latest meta three weeks ago.